[Conformance checking] More refactoring of associated type inference.
Move the rest of associated type inference into the new source file,
and split RequirementEnvironment into its own AST-level header, because it
can be re-used and has no ties to the type checker.

(cherry picked from commit 82dde17ef432378445c7d655f5a6376e60fe3dd3)
